webapp/requirements.txt
gunicorn
Flask
Jinja2
Werkzeug
mecab-python3
gensim
bert_score
fugashi
ipadic
